--- src/app/api/seller/orders/_lib/mapper.ts
@@ -1,4 +1,0 @@
-export {
-  mapOrderDetailRow,
-  mapOrderListRow,
-} from '@/app/api/orders/_lib/mapper';
